+++ gendarme/FSharpExamples/BadRecursiveInvocationRule.fs
@@ -1,0 +1,9 @@
+﻿namespace BadRecursiveInvocation
+
+type Handler() =
+  class
+    [<DefaultValue(true); System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Gendarme.Rules.Design","AvoidVisibleFieldsRule")>]
+    val mutable activeRow : int
+  end
+
+